TST: Move epics stuff into single test file
