+++ bluesky/tests/test_hardware_checklist.py
@@ -23,21 +23,3 @@ def test_connect_channelarchiver():
     # Just test failure, not success.
     with pytest.raises(RuntimeError):
         connect_channelarchiver('http://bnl.gov/asfoijewapfoia')
-
-
-def test_connect_pv():
-    try:
-        import epics
-    except ImportError as ie:
-        pytest.skip("Epics is not installed. Skipping epics test section of "
-                    "bluesky. ImportError: {}".format(ie))
-    pv_name = 'BSTEST:VAL'
-    connect_pv(pv_name)
-    epics.caput(pv_name, 5, wait=True)
-    assert_pv_equal(pv_name, 5)
-    assert_pv_greater(pv_name, 4)
-    assert_pv_less(pv_name, 6)
-    assert_pv_in_band(pv_name, 4, 6)
-    with pytest.raises(AssertionError):
-        assert_pv_in_band(pv_name, 2, 4)
-    assert_pv_out_of_band(pv_name, 2, 4)
